analysis-video/pyproject.toml
[project]
name = "analysis-video"
version = "0.1.0"
description = "MeFit 면접 영상·음성 후처리 Lambda 함수"
requires-python = ">=3.14"

[project.optional-dependencies]
dev = ["pytest", "boto3-stubs", "mypy"]
